// src/app/models/user.model.ts
export interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  address?: {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: {
      lat: string;
      lng: string;
    };
  };
  phone?: string;
  website?: string;
  company?: {
    name: string;
    catchPhrase: string;
    bs: string;
  };
  updatedAt?: string;
  syncStatus?: 'pending' | 'synced' | 'failed';
  syncError?: string;
}